+++ src/cli/flags/total_amount.rs
@@ -4,12 +4,12 @@ use clap::Parser;
 pub struct TotalAmount {
     /// Write the total amount to add
     #[clap(short, long)]
-    pub amount: Option<u64>,
+    pub amount: Option<f64>,
 }
 
 #[derive(Debug, Parser)]
 pub struct UpdateTotalAmount {
     /// Write the total amount to update
     #[clap(short, long)]
-    pub amount: u64,
+    pub amount: f64,
 }
